# src/swagger.yaml
openapi: 3.0.0
info:
  title: OKR Express API
  version: 1.0.0
  description: API for managing OKRs, companies, and teams
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OKR:
      type: object
      properties:
        _id:
          type: string
        team:
          type: string
        createdBy:
          type: string
        objective:
          type: string
        description:
          type: string
        keyResults:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              progress:
                type: number
        progress:
          type: number
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /api/teams/{teamId}/okrs:
    post:
      summary: Create OKR in team
      tags: [OKRs]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - objective
              properties:
                objective:
                  type: string
                description:
                  type: string
                keyResults:
                  type: array
                  items:
                    type: object
                    properties:
                      title:
                        type: string
                      description:
                        type: string
      responses:
        '201':
          description: OKR created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  okr:
                    $ref: '#/components/schemas/OKR'
        '401':
          description: Unauthorized
        '404':
          description: Team not found
        '403':
          description: User does not have access to this team

    get:
      summary: Get all OKRs in team
      tags: [OKRs]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of OKRs in team
          content:
            application/json:
              schema:
                type: object
                properties:
                  okrs:
                    type: array
                    items:
                      $ref: '#/components/schemas/OKR'
        '401':
          description: Unauthorized
        '404':
          description: Team not found
        '403':
          description: User does not have access to this team
  /api/okrs/{okrId}/key-results:
    post:
      summary: Add key result to OKR
      tags: [OKRs]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: okrId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Key result added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  okr:
                    $ref: '#/components/schemas/OKR'
        '401':
          description: Unauthorized
        '404':
          description: OKR not found
        '403':
          description: User does not have permission to manage this OKR
...